+++ htdocs/js/pages/admin/Groups.class.js
@@ -1386,9 +1386,9 @@ Page.Groups = class Groups extends Page.ServerUtils {
 				// self.getShortDateTime( job.started ),
 				'<div id="d_vg_jt_server_' + job.id + '">' + self.getNiceServer(job.server, true) + '</div>',
 				'<div id="d_vg_jt_state_' + job.id + '">' + self.getNiceJobState(job) + '</div>',
-				'<div id="d_vg_jt_elapsed_' + job.id + '">' + self.getNiceJobElapsedTime(job, false) + '</div>',
+				'<div id="d_vg_jt_elapsed_' + job.id + '">' + self.getNiceJobElapsedTime(job, true, false) + '</div>',
 				'<div id="d_vg_jt_progress_' + job.id + '">' + self.getNiceJobProgressBar(job) + '</div>',
-				'<div id="d_vg_jt_remaining_' + job.id + '">' + self.getNiceJobRemainingTime(job, false) + '</div>',
+				'<div id="d_vg_jt_remaining_' + job.id + '">' + self.getNiceJobRemainingTime(job, true, true) + '</div>',
 				
 				'<button class="link danger" onClick="$P().doAbortJob(\'' + job.id + '\')"><b>Abort Job</b></button>'
 			];
@@ -1634,8 +1634,8 @@ Page.Groups = class Groups extends Page.ServerUtils {
 			jobs.forEach( function(job) {
 				div.find('#d_vg_jt_state_' + job.id).html( self.getNiceJobState(job) );
 				div.find('#d_vg_jt_server_' + job.id).html( self.getNiceServer(job.server, true) );
-				div.find('#d_vg_jt_elapsed_' + job.id).html( self.getNiceJobElapsedTime(job, false) );
-				div.find('#d_vg_jt_remaining_' + job.id).html( self.getNiceJobRemainingTime(job, false) );
+				div.find('#d_vg_jt_elapsed_' + job.id).html( self.getNiceJobElapsedTime(job, true, false) );
+				div.find('#d_vg_jt_remaining_' + job.id).html( self.getNiceJobRemainingTime(job, true, true) );
 				
 				// update progress bar without redrawing it (so animation doesn't jitter)
 				self.updateJobProgressBar(job, '#d_vg_jt_progress_' + job.id + ' > div.progress_bar_container');